+++ cbftp/src/loadmonitorcallback.h
@@ -1,7 +1,0 @@
-#pragma once
-
-class LoadMonitorCallback {
-public:
-  virtual void recommendedPerformanceLevelChanged(int newlevel) = 0;
-  virtual ~LoadMonitorCallback() {}
-};