+++ cbftp/src/racestatus.h
@@ -1,8 +1,0 @@
-#pragma once
-
-enum class RaceStatus {
-  RUNNING,
-  DONE,
-  ABORTED,
-  TIMEOUT
-};