+++ cbftp/src/requestcallback.h
@@ -1,8 +1,0 @@
-#pragma once
-
-class RequestCallback {
-public:
-  virtual ~RequestCallback() = default;
-  virtual void intermediateData(void* service, int requestid, void* data) {}
-  virtual void requestReady(void* service, int requestid) = 0;
-};