+++ cbftp/src/ui/ncurseswrap.h
@@ -1,8 +1,0 @@
-#ifndef _XOPEN_SOURCE_EXTENDED
-#define _XOPEN_SOURCE_EXTENDED
-#endif
-#ifndef NO_NCURSESW
-#include <ncursesw/curses.h>
-#else
-#include <ncurses.h>
-#endif