+++ cbftp/src/ui/textwrap.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <list>
-#include <string>
-
-class TextWrap {
-public:
-  static std::list<std::string> wrap(const std::string & text, unsigned int linelen);
-};