+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
-## Asciiville-2.0.1r2 (2023-01-20)
+## Asciiville-2.0.1r2 (2023-01-24)
 
+* Move `btop` to external package install
+* Move `cbftp` to external package install
+* Move `endoh1` to external package install
 * Add `term-image` command - display and browse images in the terminal
+* Move installation location back to `/usr` except macOS to `/usr/local`
 
 ## Asciiville-2.0.1r1 (2023-01-10)